change lifetime variables to have a unification table
We still never actually **unify** them.
